+++ Properties/BuildVersion.g.cs
@@ -2,7 +2,7 @@
 {
 public static class BuildVersion
 {
-public const string Version = "1.3.0";
-public const string VersionLabel = "v1.3.0";
+public const string Version = "1.3.1";
+public const string VersionLabel = "v1.3.1";
 }
 }